ml-inference/requirements.txt
fastapi
uvicorn
torch
transformers
faiss-cpu
pydantic